Delete dialog migrated
